Started adding the service to track apps

+++ app/src/main/java/com/joshuarichardson/fivewaystowellbeing/storage/entity/PhysicalActivity.java
@@ -65,7 +65,7 @@ public class PhysicalActivity {
         return this.isPending;
     }
 
-    public boolean getTimeStatus(int endTime) {
+    public boolean getTimeStatus(long endTime) {
         int tenMins = 600000;
         return endTime >= (startTime + tenMins);
     }
